+++ SmartCursorPlus/SmartCursorModSystem.cs
@@ -1,0 +1,1 @@
+../SmartCursor/SmartCursorModSystem.cs